+++ src/full/Agda/Syntax/Notation.hs
@@ -123,11 +123,9 @@ mkNotation _ [] = throwError "empty notation is disallowed"
 mkNotation holes ids = do
   unless uniqueHoleNames     $ throwError "syntax must use unique argument names"
   let xs :: Notation = map mkPart ids
-  unless (isAlternating xs)  $ throwError $ concat
-     [ "syntax must alternate holes ("
+  unless (noAdjacentHoles xs)  $ throwError $ concat
+     [ "syntax must not contain adjacent holes ("
      , prettyHoles
-     , ") and non-holes ("
-     , prettyNonHoles xs
      , ")"
      ]
   unless (isExprLinear xs)   $ throwError "syntax must use holes exactly once"
@@ -144,16 +142,6 @@ mkNotation holes ids = do
 
       prettyHoles :: String
       prettyHoles = List.unwords $ map (rawNameToString . rangedThing) holeNames
-
-      nonHoleNames :: Notation -> [RString]
-      nonHoleNames xs = flip mapMaybe xs $ \case
-        WildHole{}   -> Just $ unranged "_"
-        IdPart x     -> Just x
-        BindHole{}   -> Nothing
-        NormalHole{} -> Nothing
-
-      prettyNonHoles :: Notation -> String
-      prettyNonHoles = List.unwords . map (rawNameToString . rangedThing) . nonHoleNames
 
       mkPart ident = maybe (IdPart ident) (`withRangeOf` ident) $ lookup ident holeMap
 
@@ -204,10 +192,11 @@ mkNotation holes ids = do
                           [ i | (i, h) <- numberedHoles,
                                 LambdaHole x _ <- [namedArg h], rangedThing x /= "_" ]
 
-      isAlternating :: [GenPart] -> Bool
-      isAlternating []       = __IMPOSSIBLE__
-      isAlternating [x]      = True
-      isAlternating (x:y:xs) = isAHole x /= isAHole y && isAlternating (y:xs)
+      noAdjacentHoles :: [GenPart] -> Bool
+      noAdjacentHoles []       = __IMPOSSIBLE__
+      noAdjacentHoles [x]      = True
+      noAdjacentHoles (x:y:xs) =
+        not (isAHole x && isAHole y) && noAdjacentHoles (y:xs)
 
       isSingleHole :: [GenPart] -> Bool
       isSingleHole = \case